Extract HIR variant constructor helpers
Move enum variant resolution and positional/named payload constructor lowering into hir/variants.rs so the HIR facade keeps shrinking behind the compiler.hir boundary.

Update the compiler source decomposition plan and ratchet ceilings for hir.rs at 6259 lines, top-file share 0.8376, top-file lines 74448, and the new variants.rs helper module.

Validation: cargo fmt --manifest-path stage1/crates/axiomc/Cargo.toml; cargo test --manifest-path stage1/Cargo.toml -p axiomc --lib hir::boundary_tests -- --nocapture; python3 scripts/ci/report-compiler-source-monoliths.py --json --check-plan --check-ratchet; make stage1-hir-boundary; make stage1-compiler-source-monoliths-test; make stage1-compiler-source-monoliths; cargo test --manifest-path stage1/Cargo.toml -p axiomc --lib; git diff --check.
